% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{callplot}
\alias{callplot}
\title{Mid-level plotting utility function to execute plotting statements}
\usage{
callplot(x)
}
\arguments{
\item{x}{The object to be plotted.}
}
\description{
The function calls a plot which is stored either as an expression, a function, or an R script.
}
