% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_dirs.R
\name{createDSProject}
\alias{createDSProject}
\title{Create directory structure for a data science project}
\usage{
createDSProject(pname = NULL, path = ".")
}
\arguments{
\item{pname}{a character, stating the name of the project}

\item{path}{a character, specifying the location of the project}
}
\description{
Initialise a data science project. This function creates a directory
structure to help you organise your project. The structure is inspired from
JEFFERY LEEKS TUTORIAL. Along with the directory structure, it creates a
README file in Rmarkdown which can then be used to produce a GitHub
compatible markdown document. It also creates a plot diary to help you
organise plots generated during the project.
}
\details{
The README.Rmd file is created in the project directory and the
plot_diary.Rmd created in the auto generated figures folder
}
\examples{
\dontrun{
# create project named "ds_project" in home folder
createDSProject('ds_project', '~')
}

}
\seealso{
\code{\link{updateReadme}}
}
