% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscapeAccessoryFunctions.R
\name{getCytoscapeLayout}
\alias{getCytoscapeLayout}
\title{Get layout for a cytoscape network}
\usage{
getCytoscapeLayout(normalise = TRUE)
}
\arguments{
\item{normalise}{a logical, indicating whether coordinates should be
normalised using the z-transform or not.}
}
\value{
a data.frame with the x and y coordinates of nodes. Node names are
  stored as rownames of the data.frame.
}
\description{
This function retrieves the layout of an active cytoscape network using the
`RCy3` package. The layout is retrieved as node coordinates.
}
\examples{
# getCytoscapeLayout()

}
