% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{getOS}
\alias{getOS}
\title{Get operating system architecture}
\usage{
getOS()
}
\description{
This package returns the type of OS you use.
}
\details{
This function is stolen from somewhere on the web. I will look it up later :).
}
