% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{openimage}
\alias{openimage}
\title{Function to open graphical devices of all kind}
\usage{
openimage(name = NULL, path = "", format = NULL, width = NULL,
  height = NULL)
}
\arguments{
\item{name}{Character string. The name of the file without the extension. Leaving this the default \code{NULL} will open an x11 window.}

\item{path}{Character strings. The path to the file that you want. I decided to keep this separate as many times I use separater variables to indicate folder structure.}

\item{format}{Character string. The used device/file extension. If left as \code{NULL}, then the function will open an x11 window.}

\item{width}{The width of the plot}

\item{height}{The height of the plot.}
}
\description{
This wrapper function will open image devices, without having to call different functions.
}
