% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{pan11}
\alias{pan11}
\alias{pan21}
\title{Predefined multipanel plotting.}
\usage{
pan11(a, name = NULL, format = NULL, path = NULL, width = NULL,
  height = NULL, pind = "small")

pan21(a, b, name = NULL, format = NULL, path = NULL, width = NULL,
  height = NULL, pind = c("A", "B"), outmai = NULL, intop = NULL,
  inbot = NULL)
}
\arguments{
\item{a}{Expression or path (as character string). Panel A plotting insturctions.}

\item{format}{The extension of the final image/the used plotting device.}

\item{path}{Character string, the path in which the final plot should be placed.}

\item{width}{The width of the plot.}

\item{height}{Integer vector. The heights of the plots in inches.}

\item{pind}{Character vector of the panel identifiers.}

\item{b}{Expression or path (as character string). Panel B plotting insturctions.}

\item{outmai}{Numeric vector, the outer margins of the plot in inches (as in par()$mai).}

\item{intop}{Numeric value. The top margin of the lower plot in inches. If outmai is NULL then the top margin of the plot.}

\item{inbot}{Numeric value. The bottom marign of the upper plot.}
}
\description{
The function is useful for the effective combination of multipanel plots.
}
