% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figs.R
\name{panelID}
\alias{panelID}
\title{Put a panel indicator on an R plot}
\usage{
panelID(label, location = "topleft", x = NULL, y = NULL,
  offset = c(0, 0), ...)
}
\arguments{
\item{label}{Character string. The label you want to put on the panel.}

\item{location}{Character string, similar to that of legend().}

\item{x}{Numeric, the x parameter of the label.}

\item{y}{Numeric, the y parameter of the label.}

\item{offset}{offset from the calculated coordinates.}

\item{...}{arguments passed to the text() function.}
}
\description{
Panel indicator function.
}
\details{
The function is stolen from:
https://waterprogramming.wordpress.com/2015/12/02/easy-labels-for-multi-panel-plots-in-r/
with minor modifications.
}
