% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDimReduct.R
\name{plotPCA}
\alias{plotPCA}
\alias{plotPCA,DGEList-method}
\alias{plotPCA,ExpressionSet-method}
\alias{plotPCA,SummarizedExperiment-method}
\alias{plotPCA,SingleCellExperiment-method}
\alias{plotPCA,SpatialExperiment-method}
\title{Compute and plot the results of a PCA analysis on gene expression data}
\usage{
plotPCA(edata, dims = c(1, 2), ...)

\S4method{plotPCA}{DGEList}(edata, dims, precomputed = NULL, rl = 1, subset_row = NULL, ...)

\S4method{plotPCA}{ExpressionSet}(edata, dims, precomputed = NULL, rl = 1, subset_row = NULL, ...)

\S4method{plotPCA}{SummarizedExperiment}(
  edata,
  dims,
  assay = 1,
  precomputed = NULL,
  rl = 1,
  subset_row = NULL,
  ...
)

\S4method{plotPCA}{SingleCellExperiment}(edata, dims, dimred = "PCA", rl = 1, ...)

\S4method{plotPCA}{SpatialExperiment}(edata, dims, dimred = "PCA", rl = 1, ...)
}
\arguments{
\item{edata}{a DGEList, SummarizedExperiment or ExpressionSet object
containing gene expression data.}

\item{dims}{a numeric, containing 2 values specifying the dimensions to plot.}

\item{...}{aesthetic mappings to pass to `ggplot2::aes_string()`.}

\item{precomputed}{a dimensional reduction results from `stats::prcomp`.}

\item{rl}{a numeric, specifying the relative scale factor to apply to text on
the plot.}

\item{subset_row}{NULL (default) or character, for the features to use in
calculating the PCA with `scater::calculatePCA`. The default will use the
top 500 features.}

\item{assay}{a numeric or character, specifying the assay to use (for
SummarizedExperiment).}

\item{dimred}{a string or integer scalar indicating the reduced dimension
result in `reducedDims(object)` to plot.}
}
\value{
a ggplot2 object
}
\description{
Compute and plot the results of a PCA analysis on gene expression data
}
\examples{
se = emtdata::cursons2018_se()
dge = emtdata::asDGEList(se)
plotPCA(dge, colour = Subline)

}
