% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRLE.R
\name{plotRLE}
\alias{plotRLE}
\alias{plotRLE,DGEList-method}
\alias{plotRLE,ExpressionSet-method}
\alias{plotRLE,SummarizedExperiment-method}
\title{Compute and plot relative log expression (RLE) values of gene expression data}
\usage{
plotRLE(edata, ordannots = c(), rl = 1, ...)

\S4method{plotRLE}{DGEList}(edata, ordannots = c(), rl = 1, ...)

\S4method{plotRLE}{ExpressionSet}(edata, ordannots = c(), rl = 1, ...)

\S4method{plotRLE}{SummarizedExperiment}(edata, ordannots = c(), rl = 1, ...)
}
\arguments{
\item{edata}{a DGEList, SummarizedExperiment or ExpressionSet object
containing gene expression data.}

\item{ordannots}{a character, specifying the annotations to use to order
samples. Samples are ordered successively using each annotation. If the
columns to use are "A" and "B" (from the annotation table), samples are
first ordered using "A" and then "B" within each category of "A".}

\item{rl}{a numeric, specifying the relative scale factor to apply to text on
the plot.}

\item{...}{aesthetic mappings to pass to `ggplot2::aes_string()`.}
}
\value{
a ggplot2 object, containing the RLE plot.
}
\description{
Compute and plot relative log expression (RLE) values of gene expression data
}
\examples{
se = emtdata::cursons2018_se()
dge = emtdata::asDGEList(se)
plotRLE(dge, colour = Subline, lty = Treatment, lwd = 2, ordannots =
c(Subline, Treatment))

}
