% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatial.R
\name{plotSpotAnnotation}
\alias{plotSpotAnnotation}
\title{Plot spatial transcriptomic annotations per spot}
\usage{
plotSpotAnnotation(spe, rl = 1, ...)
}
\arguments{
\item{spe}{a SpatialExperiment or SpatialExperimentList object.}

\item{rl}{a numeric, specifying the relative scale factor to apply to text on
the plot.}

\item{...}{aesthetic mappings to pass to `ggplot2::aes_string()`.}
}
\value{
a ggplot2 object
}
\description{
Plot spatial transcriptomic annotations per spot
}
