% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatial.R
\name{plotSpotExpression}
\alias{plotSpotExpression}
\title{Plot spatial transcriptomic gene expression per spot}
\usage{
plotSpotExpression(spe, assay, rl = 1, ...)
}
\arguments{
\item{spe}{a SpatialExperiment or SpatialExperimentList object.}

\item{assay}{a numeric or character, specifying the assay to use (for
SummarizedExperiment).}

\item{rl}{a numeric, specifying the relative scale factor to apply to text on
the plot.}

\item{...}{aesthetic mappings to pass to `ggplot2::aes_string()`.}
}
\value{
a ggplot2 object
}
\description{
Plot spatial transcriptomic gene expression per spot
}
