% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatial.R
\name{plotSpotReducedDim}
\alias{plotSpotReducedDim}
\title{Plot spatial transcriptomic reduced dimension per spot}
\usage{
plotSpotReducedDim(
  spe,
  dimred = SingleCellExperiment::reducedDimNames(spe),
  rl = 1,
  ...
)
}
\arguments{
\item{spe}{a SpatialExperiment or SpatialExperimentList object.}

\item{dimred}{a string or integer scalar indicating the reduced dimension
result in reducedDims(object) to plot.}

\item{rl}{a numeric, specifying the relative scale factor to apply to text on
the plot.}

\item{...}{aesthetic mappings to pass to `ggplot2::aes_string()`.}
}
\value{
a ggplot2 object
}
\description{
Plot spatial transcriptomic reduced dimension per spot
}
