% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpatial.R
\name{plotSpots}
\alias{plotSpots}
\title{Plot spatial transcriptomic annotations per spot}
\usage{
plotSpots(
  spe,
  what = c("annotation", "expression", "reduceddim"),
  ...,
  assay = SummarizedExperiment::assayNames(spe),
  dimred = SingleCellExperiment::reducedDimNames(spe),
  img = TRUE,
  crop = FALSE,
  imgAlpha = 1,
  rl = 1,
  circles = FALSE
)
}
\arguments{
\item{spe}{a SpatialExperiment object.}

\item{what}{a character specifying what aspect should be plot, "annotation",
"expression" or reduced dimension ("reduceddim").}

\item{...}{aesthetic mappings to pass to `ggplot2::aes_string()`.}

\item{assay}{a numeric or character, specifying the assay to use (for
SummarizedExperiment).}

\item{dimred}{a string or integer scalar indicating the reduced dimension
result in `reducedDims(object)` to plot.}

\item{img}{a logical indicating whether the tissue image (if present) should}

\item{rl}{a numeric, specifying the relative scale factor to apply to text on
the plot.}
}
\value{
a ggplot2 object
}
\description{
Plot spatial transcriptomic annotations per spot
}
