% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAlignmentResults.R
\name{readFeatureCountsRes}
\alias{readFeatureCountsRes}
\title{Read results of a Subread analysis into a DGEList object}
\usage{
readFeatureCountsRes(fpath, plot = TRUE)
}
\arguments{
\item{fpath}{a character, specifying the path of the analysis (ideally using
bhuva.d's scripts).}

\item{plot}{a logical, stating whether to plot summary statistics for the
alignment or not.}
}
\value{
a DGEList object, containing transcriptomics data
}
\description{
Creates a DGEList object using files from a Subread analysis. The function
will look for a file called 'SraRunTable.txt' for sample annotations, a
'*_counts.txt' file(s) for expression data and 'counts.txt.summary' for
summary statistics (optional).
}
