% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_readme.R
\name{updateReadme}
\alias{updateReadme}
\title{Update the README file for the project}
\usage{
updateReadme(dsproj = ".")
}
\arguments{
\item{dsproj}{a character, providing the path of the project where the file
needs to be updated.}
}
\description{
Scans the project directory provided and creates entries for files in the
README.Rmd file. Descriptions of files can then be modified manually.
}
\details{
Only the description of files/directories should be modified in the
README.Rmd file. Modifying the comment tags or adding/removing text from the
file names will result in their removal in the next update.
}
\examples{
\dontrun{
# for "ds_project" in home folder
updateReadme('~/ds_project')
}

}
