% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{setver}
\alias{setver}
\alias{getver}
\title{Assign a new value to the session version}
\usage{
setver(ver)

getver()
}
\arguments{
\item{ver}{Character. The session version identifier.}
}
\description{
The function is used to change the 'sessionVersion' object
}
