% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R, R/utility.R
\name{newstudy}
\alias{newstudy}
\title{Function to create standard folder structure of a new study}
\usage{
newstudy(ver = NULL)

newstudy(ver = NULL)
}
\arguments{
\item{ver}{Character. The session version identifier.}

\item{ver}{Character. The session version identifier.}
}
\description{
New folder structure

New folder structure
}
